requirements.txt
py-cord
jishaku
requests
Pillow